feat: move COUNT_OF_DATES_FOR_CHART to config/charts
