+++ src/config/general.ts
@@ -1,8 +1,3 @@
 const ACCOUNT_ID_TYPE_NAME = 'AccountId';
 
-// Get 6 values to be able to calculate difference between 5 days ago and 6 days ago.
-// Thus volume per day 5 days ago can be displayed.
-// Cumulative volume is also only displayed to 5 days.
-const COUNT_OF_DATES_FOR_CHART = 6;
-
-export { ACCOUNT_ID_TYPE_NAME, COUNT_OF_DATES_FOR_CHART };
+export { ACCOUNT_ID_TYPE_NAME };
